requirements.txt
asgiref==3.3.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
google-api-core==1.26.0
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
protobuf==3.15.3
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.0.1
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytz==2021.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
schema==0.7.4
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
whitenoise==5.2.0
